+++ src/test/kotlin/org/drexa1/gedcom/TestGedcomLexer.kt
@@ -1,7 +1,7 @@
 package org.drexa1.gedcom
 
 import org.drexa1.gedcom.highlighter.GedcomLexer
-import java.io.File
+import org.drexa1.gedcom.highlighter.GedcomTokens
 import kotlin.test.Test
 
 @kotlin.concurrent.atomics.ExperimentalAtomicApi
@@ -12,25 +12,25 @@ class TestGedcomLexer {
         // GIVEN
         val resource = object {}.javaClass.classLoader.getResource("sample.ged")
             ?: throw IllegalArgumentException("sample.ged not found")
-        val gedcomText = File(resource.toURI()).readText()
+        val gedcomText = resource.readText()
         val lexer = GedcomLexer()
         lexer.start(gedcomText)
-        // GIVEN
+        // WHEN
         while (true) {
             val tokenType = lexer.tokenType ?: break
             val tokenText = lexer.bufferSequence.subSequence(lexer.tokenStart, lexer.tokenEnd).toString()
-            // THEN
             when (tokenType) {
-                GedcomTokens.LEVEL -> assert(tokenText.toIntOrNull() != null)
-                GedcomTokens.AT -> assert(tokenText == "@")
-                GedcomTokens.POINTER -> assert(tokenText.isNotEmpty())
-                GedcomTokens.INDI_POINTER -> assert(tokenText.startsWith('I'))
-                GedcomTokens.FAM_POINTER -> assert(tokenText.startsWith('F'))
-                GedcomTokens.TAG -> assert(tokenText.isNotEmpty())
-                GedcomTokens.VALUE -> assert(tokenText.isNotEmpty())
-                GedcomTokens.WHITESPACE -> assert(tokenText.single().isWhitespace())
+                // THEN
+                GedcomTokens.LEVEL           -> assert(tokenText.toIntOrNull() != null)
+                GedcomTokens.AT              -> assert(tokenText == "@")
+                GedcomTokens.POINTER         -> assert(tokenText.isNotEmpty())
+                GedcomTokens.INDI_POINTER    -> assert(tokenText.startsWith('I'))
+                GedcomTokens.FAM_POINTER     -> assert(tokenText.startsWith('F'))
+                GedcomTokens.TAG             -> assert(tokenText.isNotEmpty())
+                GedcomTokens.VALUE           -> assert(tokenText.isNotEmpty())
+                GedcomTokens.WHITESPACE      -> assert(tokenText.single().isWhitespace())
                 GedcomTokens.NEW_LINE_INDENT -> assert(tokenText.single().code == 10)
-                GedcomTokens.BAD_CHARACTER -> assert(tokenText.length == 1)
+                GedcomTokens.BAD_CHARACTER   -> assert(tokenText.length == 1)
             }
             lexer.advance()
         }