+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [1.3.0] - Aug 24, 2025
+### Added
+- Improved lexer to support META, INDI & FAM code sections
+- Added ColorSettingsPage
+- Improved color schema suiting light and dark themes
+
 ## [1.2.0] - Aug 22, 2025
 ### Added
 - Flex & Antlr lexers